Move proto-serialization tests into programs_test.py
