+++ cirq/google/xmon_gates_test.py
@@ -12,30 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pytest
 import numpy as np
 
 import cirq
 import cirq.google as cg
-
-
-def assert_proto_dict_convert(gate_cls, gate, proto_dict, *qubits):
-    from cirq.google.programs import _gate_to_proto_dict
-    assert _gate_to_proto_dict(gate, qubits) == proto_dict
-    assert gate_cls.from_proto_dict(proto_dict) == gate(*qubits)
-
-
-def test_parameterized_value_from_proto():
-    from_proto = cg.XmonGate.parameterized_value_from_proto_dict
-
-    m1 = {'raw': 5}
-    assert from_proto(m1) == 5
-
-    with pytest.raises(ValueError):
-        from_proto({})
-
-    m3 = {'parameter_key': 'rr'}
-    assert from_proto(m3) == cirq.Symbol('rr')
 
 
 def test_measurement_eq():
@@ -47,104 +27,6 @@ def test_measurement_eq():
                                                           invert_mask=(True,)))
     eq.make_equality_group(lambda: cg.XmonMeasurementGate(key='',
                                                           invert_mask=(False,)))
-
-
-def test_single_qubit_measurement_proto_dict_convert():
-    gate = cg.XmonMeasurementGate('test')
-    proto_dict = {
-        'measurement': {
-            'targets': [
-                {
-                    'row': 2,
-                    'col': 3
-                }
-            ],
-            'key': 'test'
-        }
-    }
-    assert_proto_dict_convert(cg.XmonMeasurementGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-
-def test_single_qubit_measurement_invalid_dict():
-    proto_dict = {
-        'measurement': {
-            'targets': [
-                {
-                    'row': 2,
-                    'col': 3
-                }
-            ],
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.XmonMeasurementGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'measurement': {
-            'targets': [
-                {
-                    'row': 2,
-                    'col': 3
-                }
-            ],
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.XmonMeasurementGate.from_proto_dict(proto_dict)
-
-
-def test_single_qubit_measurement_to_proto_dict_convert_invert_mask():
-    gate = cg.XmonMeasurementGate('test', invert_mask=(True,))
-    proto_dict = {
-        'measurement': {
-            'targets': [
-                {
-                    'row': 2,
-                    'col': 3
-                }
-            ],
-            'key': 'test',
-            'invert_mask': ['true']
-        }
-    }
-    assert_proto_dict_convert(cg.XmonMeasurementGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-
-def test_multi_qubit_measurement_to_proto_dict():
-    gate = cg.XmonMeasurementGate('test')
-    proto_dict = {
-        'measurement': {
-            'targets': [
-                {
-                    'row': 2,
-                    'col': 3
-                },
-                {
-                    'row': 3,
-                    'col': 4
-                }
-            ],
-            'key': 'test'
-        }
-    }
-    assert_proto_dict_convert(cg.XmonMeasurementGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3), cirq.GridQubit(3, 4))
-
-
-@cirq.testing.only_test_in_python3
-def test_measurement_repr():
-    gate = cg.XmonMeasurementGate('test', invert_mask=(True,))
-    assert repr(gate) == 'XmonMeasurementGate(\'test\', (True,))'
-
-
-def test_invalid_measurement_gate():
-    with pytest.raises(ValueError, match='length'):
-        cg.XmonMeasurementGate('test', invert_mask=(True,)).to_proto_dict(
-            cirq.GridQubit(2, 3), cirq.GridQubit(3, 4))
-    with pytest.raises(ValueError, match='no qubits'):
-        cg.XmonMeasurementGate('test').to_proto_dict()
 
 
 def test_z_eq():
@@ -163,59 +45,10 @@ def test_z_eq():
         cg.ExpZGate(half_turns=10.5))
 
 
-def test_z_proto_dict_convert():
-    gate = cg.ExpZGate(half_turns=cirq.Symbol('k'))
-    proto_dict = {
-        'exp_z': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.ExpZGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-    gate = cg.ExpZGate(half_turns=0.5)
-    proto_dict = {
-        'exp_z': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'half_turns': {
-                'raw': 0.5
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.ExpZGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-
-def test_z_invalid_dict():
-    proto_dict = {
-        'exp_z': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.ExpZGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'exp_z': {
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.ExpZGate.from_proto_dict(proto_dict)
+@cirq.testing.only_test_in_python3
+def test_measurement_repr():
+    gate = cg.XmonMeasurementGate('test', invert_mask=(True,))
+    assert repr(gate) == "XmonMeasurementGate('test', (True,))"
 
 
 def test_z_matrix():
@@ -241,91 +74,6 @@ def test_z_parameterize():
 def test_z_repr():
     gate = cg.ExpZGate(half_turns=0.25)
     assert repr(gate) == 'ExpZGate(half_turns=0.25)'
-
-
-def test_cz_proto_dict_convert():
-    gate = cirq.CZ**cirq.Symbol('k')
-    proto_dict = {
-        'exp_11': {
-            'target1': {
-                'row': 2,
-                'col': 3
-            },
-            'target2': {
-                'row': 3,
-                'col': 4
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.XmonGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3), cirq.GridQubit(3, 4))
-
-    gate = cirq.CZ**0.5
-    proto_dict = {
-        'exp_11': {
-            'target1': {
-                'row': 2,
-                'col': 3
-            },
-            'target2': {
-                'row': 3,
-                'col': 4
-            },
-            'half_turns': {
-                'raw': 0.5
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.XmonGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3), cirq.GridQubit(3, 4))
-
-
-def test_cz_invalid_dict():
-    proto_dict = {
-        'exp_11': {
-            'target2': {
-                'row': 3,
-                'col': 4
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    with pytest.raises(ValueError, match='missing required fields'):
-        cg.XmonGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'exp_11': {
-            'target1': {
-                'row': 2,
-                'col': 3
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    with pytest.raises(ValueError, match='missing required fields'):
-        cg.XmonGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'exp_11': {
-            'target1': {
-                'row': 2,
-                'col': 3
-            },
-            'target2': {
-                'row': 3,
-                'col': 4
-            },
-        }
-    }
-    with pytest.raises(ValueError, match='missing required fields'):
-        cg.XmonGate.from_proto_dict(proto_dict)
 
 
 def test_w_eq():
@@ -377,87 +125,6 @@ def test_w_str():
     assert str(cg.ExpWGate(axis_half_turns=0.5, half_turns=0.25)) == 'Y^0.25'
     assert str(cg.ExpWGate(axis_half_turns=0.25,
                            half_turns=0.5)) == 'W(0.25)^0.5'
-
-
-def test_w_to_proto_dict():
-    gate = cg.ExpWGate(half_turns=cirq.Symbol('k'), axis_half_turns=1)
-    proto_dict = {
-        'exp_w': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'axis_half_turns': {
-                'raw': 1
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.ExpWGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-    gate = cg.ExpWGate(half_turns=0.5, axis_half_turns=cirq.Symbol('j'))
-    proto_dict = {
-        'exp_w': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'axis_half_turns': {
-                'parameter_key': 'j'
-            },
-            'half_turns': {
-                'raw': 0.5
-            }
-        }
-    }
-    assert_proto_dict_convert(cg.ExpWGate, gate, proto_dict,
-                              cirq.GridQubit(2, 3))
-
-
-def test_w_invalid_dict():
-    proto_dict = {
-        'exp_w': {
-            'axis_half_turns': {
-                'raw': 1
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.ExpWGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'exp_w': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'half_turns': {
-                'parameter_key': 'k'
-            }
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.ExpWGate.from_proto_dict(proto_dict)
-
-    proto_dict = {
-        'exp_w': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-            'axis_half_turns': {
-                'raw': 1
-            },
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.ExpWGate.from_proto_dict(proto_dict)
 
 
 def test_w_decomposition():
@@ -518,31 +185,6 @@ def test_measure_key_on():
     assert cg.XmonMeasurementGate(key='a').on(q) == cirq.GateOperation(
         gate=cg.XmonMeasurementGate(key='a'),
         qubits=(q,))
-
-
-def test_unsupported_op():
-    proto_dict = {
-        'not_a_gate': {
-            'target': {
-                'row': 2,
-                'col': 3
-            },
-        }
-    }
-    with pytest.raises(ValueError):
-        cg.XmonGate.from_proto_dict(proto_dict)
-
-
-def test_invalid_to_proto_dict_qubit_number():
-    from cirq.google.programs import _gate_to_proto_dict
-    with pytest.raises(ValueError, match='Wrong number of qubits'):
-        _ = _gate_to_proto_dict(cirq.CZ**0.5, (cirq.GridQubit(2, 3),))
-    with pytest.raises(ValueError, match='Wrong number of qubits'):
-        cg.ExpZGate(half_turns=0.5).to_proto_dict(cirq.GridQubit(2, 3),
-                                                  cirq.GridQubit(3, 4))
-    with pytest.raises(ValueError, match='Wrong number of qubits'):
-        cg.ExpWGate(half_turns=0.5, axis_half_turns=0).to_proto_dict(
-            cirq.GridQubit(2, 3), cirq.GridQubit(3, 4))
 
 
 def test_cirq_symbol_diagrams():
